rename remediation and rules folders
